+++ src/delay/functions/index.ts
@@ -1,1 +1,2 @@
 export { fixedDelay } from "./fixedDelay";
+export { fixedRepeater } from "./fixedRepeater";