+++ src/delay/types/fixedRepeaterProps.ts
@@ -1,0 +1,5 @@
+export type FixedRepeaterProps = {
+  callback: () => void;
+  ms: number;
+  times: number;
+};