+++ src/delay/types/index.ts
@@ -1,1 +1,2 @@
 export type { FixedDelayProps } from "./fixedDelayProps";
+export type { FixedRepeaterProps } from "./fixedRepeaterProps";